user can update and delete stat they create

+++ app/controllers/game_stats_controller.rb
@@ -13,37 +13,33 @@ class GameStatsController < ApplicationController
             render json: {error: stat.errors.full_messages}, status: :unprocessable_entity
         end
     end 
-
+    
     def destroy 
-        stat = GameStat.find_by(id: params[:id])
-        # if @game_stat
-        #   if @reservation.user_id == @current_user.id
+        if stat = GameStat.find_by(id: params[:id])
+         if stat.game.away_team_id || stat.game.home_team_id == @current_user.team.id
             stat.destroy 
             head :no_content
-        #   else 
-        #     render json: { error: "Unauthorized: Cannot delete reservation" }, status: :unauthorized
-        #   end 
-        # else 
-        #   render json: { error: "Stat not found" }, status: :not_found
-        # end 
+          else 
+            render json: { error: "Unauthorized: Cannot delete reservation" }, status: :unauthorized
+          end 
+        else 
+          render json: { error: "Stat not found" }, status: :not_found
+        end 
       end
       def update
-        stat = GameStat.find_by(id: params[:id])
-        stat.update(stat_params)
-        render json: stat, status: :accepted
-        # if @reservation
-        #   if @reservation.user_id == @current_user.id
-        #     if @reservation.update(reservation_params)
-        #       render json: @reservation, status: :accepted
-        #     else
-        #       render json: { error: @reservation.errors.full_messages }, status: :unprocessable_entity
-        #     end
-        #   else
-        #     render json: { error: "Unauthorized: Cannot update reservation" }, status: :unauthorized
-        #   end
-        # else
-        #   render json: { error: "Reservation not found" }, status: :not_found
-        # end
+        if stat = GameStat.find_by(id: params[:id])
+             if stat.game.away_team_id || stat.game.home_team_id == @current_user.team.id
+                if stat.update(stat_params)
+                    render json: stat, status: :accepted
+                else
+                    render json: { error: stat.errors.full_messages }, status: :unprocessable_entity
+                end
+          else
+            render json: { error: "Unauthorized: Cannot update reservation" }, status: :unauthorized
+          end
+        else
+          render json: { error: "Reservation not found" }, status: :not_found
+        end
       end
 
     private
@@ -52,9 +48,7 @@ def stat_params
     params.permit(:game_id, :player_id, :played, :batting_average, :at_bat, :hits, :runs, :RBI, :stolen_base, :field_error, :fielding_percentage, :innings_pitched, :ERA, :K )
 end
 
-# def find_stat
-#     @game_stat = GameStat.find_by(id: params[:id])
-#   end
 end
+
